Add files via upload
This is a project I did recently for Geo-med displaying knowledge of Excel. I created these from originals called "Sales Data" and "Cost" to get some insights from the data pertaining to pricing of the products, this was using power query and pivot tables/charts. I was then Instructed to redo these projects on R and Python. The R project is called Geomed-2.pdf, while the python one I wrote a report for as I write my python in pycharm this is titled "Report 3 Geomed.docx